requirements.txt
python-telegram-bot
openai
dotenv
google-genai